Fully documented days controller's endpoints.
Endpoints for days controller are now fully documented.
There is a not very pretty feature in days index endpoint examples,
because no_days example is first, and it might be a little confusing,
but when days example was first the no_days example wasn't working correctly,
and I think this problem is not worth spending more time on it
so I leave it like this.